docs: use JSDocs syntax for helpers comments
